requirements/dev.txt
-r base.txt
-r fork.txt

black==23.3.0
django-debug-toolbar==4.1.0
factory-boy==3.2.1
Faker==18.10.1
flake8==6.0.0
freezegun==1.2.2
isort==5.12.0
psycopg2-binary==2.9.6
pytest==7.3.2
pytest-cov==4.1.0
pytest-django==4.5.2
pytest-factoryboy==2.5.1
pytest-mock==3.14.0
